refactor(ses): extract SesReceiptRuleService from SesService (strangler step 1)
First step of the store-based domain split. Moves the
receipt-rule-set domain — its store, lock, name validation, and the six
Create/Describe/List/Delete/SetActive/DescribeActive methods — into a new
@ApplicationScoped SesReceiptRuleService. SesService becomes a facade that
delegates, so the v1 SesQueryHandler call sites are unchanged. The domain
service is now unit-testable with just its own store and a clock.
