+++ src/test/java/io/github/hectorvent/floci/services/ses/SesServiceSuppressionLegacyKeyTest.java
@@ -58,7 +58,7 @@ class SesServiceSuppressionLegacyKeyTest {
                 new InMemoryStorage<String, ContactList>(),
                 new InMemoryStorage<String, Contact>(),
                 new InMemoryStorage<String, String>(),
-                new InMemoryStorage<String, ReceiptRuleSet>(),
+                new SesReceiptRuleService(new InMemoryStorage<String, ReceiptRuleSet>(), Clock.systemUTC()),
                 new InMemoryStorage<String, CustomVerificationEmailTemplate>(),
                 mock(SmtpRelay.class),
                 new ObjectMapper(),
